feat: removed timer code from bottom bar container
